Refactor copy link
Refactored the copy link by removing the p and making it a button adding styles and an alert on click allowing for better UX this will allow the user to click the button more easily by being able to see the style change on hover and having a larger click area adding an alet will let the user know what has happend when the button was clicked.
